Remove js-sys dep for wasm-wasi target
In open-telemetry/opentelemetry-rust#365, support was added for running
opentelemetry-jaegar as a collector under wasm. This change added many
depedencies on wasm-bindgen and js-sys. These dependencies are to all
interoperability with javascript code, but error when used in a
wasm-wasi runtime environment. Although the deeper changes were in
opentelemetry-jaeger, the base opentelemetry crate also got a js-sys
dependency in order to access SystemTime::now()

This commit modifies the `target_arch = "wasm32"` switches for the js-sys
dependency and crate::time to `all(target_arch = "wasm32", target_os =
"emscripten")` so that wasm-wasi targets not destined for running in the
browser will use std::time::SystemTime